Virtual Box 6.1 + Vagrant with nesting vm on Intel system is a great thing
Vagrantfile
...

   config.vm.provider "virtualbox" do |vb|
       vb.customize ["modifyvm", :id, "--nested-hw-virt", "on"]
       ...

+++ osp-tutorial/baseProvision.sh
@@ -4,8 +4,14 @@
 # set -o xtrace
 set -o errexit
 
-# commented out the line below, because qemu in virtual box will never make it
-## [ -e /dev/kvm ] || { echo "PROBLEM, you need to ensure nesting is enabled"; exit 1; }
+# NOTE: Assuming we are in an intel based system...
+[ -e /dev/kvm ] || { echo "PROBLEM, you need to ensure hv can nest"; exit 1; }
+grep -q Y /sys/module/kvm_intel/parameters/nested || {
+  sudo rmmod kvm-intel
+  sudo sh -c "echo 'options kvm-intel nested=y' >> /etc/modprobe.d/dist.conf"
+  sudo modprobe kvm-intel
+}
+modinfo kvm_intel | grep -q 'nested:bool' || { echo "PROBLEM, nesting did not enable"; exit 1; }
 
 yum install -y git emacs vim wget tmux
 yum install -y python3 python3-devel
